+++ h2quic/response_writer_test.go
@@ -2,6 +2,7 @@ package h2quic
 
 import (
 	"bytes"
+	"encoding/binary"
 	"io"
 	"net/http"
 	"sync"
@@ -10,6 +11,7 @@ import (
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/hpack"
 
+	quic "github.com/lucas-clemente/quic-go"
 	"github.com/lucas-clemente/quic-go/protocol"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -56,12 +58,18 @@ var _ = Describe("Response Writer", func() {
 		w            *responseWriter
 		headerStream *mockStream
 		dataStream   *mockStream
+		session      *mockSession
+		requestHost  string
 	)
-
 	BeforeEach(func() {
+		requestHost = "www.request.com"
 		headerStream = &mockStream{}
+		headerStream.id = protocol.StreamID(3)
 		dataStream = &mockStream{}
-		w = newResponseWriter(headerStream, &sync.Mutex{}, dataStream, 5)
+		dataStream.id = protocol.StreamID(5)
+		session = &mockSession{}
+		handlerFunc := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})
+		w = newResponseWriter(headerStream, &sync.Mutex{}, dataStream, dataStream.id, session, handlerFunc, requestHost)
 	})
 
 	decodeHeaderFields := func() map[string][]string {
@@ -74,7 +82,7 @@ var _ = Describe("Response Writer", func() {
 		Expect(frame).To(BeAssignableToTypeOf(&http2.HeadersFrame{}))
 		hframe := frame.(*http2.HeadersFrame)
 		mhframe := &http2.MetaHeadersFrame{HeadersFrame: hframe}
-		Expect(mhframe.StreamID).To(BeEquivalentTo(5))
+		Expect(mhframe.StreamID).To(BeEquivalentTo(dataStream.id))
 		mhframe.Fields, err = decoder.DecodeFull(hframe.HeaderBlockFragment())
 		Expect(err).ToNot(HaveOccurred())
 		for _, p := range mhframe.Fields {
@@ -148,4 +156,119 @@ var _ = Describe("Response Writer", func() {
 		Expect(err).To(MatchError(http.ErrBodyNotAllowed))
 		Expect(dataStream.dataWritten.Bytes()).To(HaveLen(0))
 	})
+
+	Context("Server push", func() {
+		var (
+			method       string
+			fakePushData string
+			pushStream   *mockStream
+			opts         *http.PushOptions
+			pushTarget   string
+		)
+		BeforeEach(func() {
+			// test that we implement http.Pusher
+			var _ http.Pusher = &responseWriter{}
+			method = "GET"
+			fakePushData = "Pushed something"
+			pushTarget = "/push_example"
+			// HandlerFunc for pusher
+			handlerFunc := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+				defer GinkgoRecover()
+				Expect(r.Method).To(Equal(method))
+				n, err := w.Write([]byte(fakePushData))
+				Expect(err).ToNot(HaveOccurred())
+				Expect(n).To(Equal(len(fakePushData)))
+			})
+			w = newResponseWriter(headerStream, &sync.Mutex{}, dataStream, dataStream.id, session, handlerFunc, requestHost)
+			opts = &http.PushOptions{
+				Method: method,
+				Header: http.Header{},
+			}
+			// add stream to open to the session so we can push:
+			pushStream = newMockStream(2)
+			session.streamsToOpen = []quic.Stream{pushStream}
+		})
+
+		It("sends a valid PUSH_PROMISE frame", func() {
+			err := w.Push(pushTarget, opts)
+			Expect(err).ToNot(HaveOccurred())
+			// Check headerStream for push promise
+			headerFrame := headerStream.dataWritten.Bytes()
+			indexStart := 3            // skip length (3 bytes)
+			indexEnd := indexStart + 1 // frame type is 1 byte
+			frameType := http2.FrameType(headerFrame[indexStart:indexEnd][0])
+			Expect(frameType).To(Equal(http2.FramePushPromise))
+			// headerStream ID
+			indexStart = indexEnd + 1 // skip flags (1 byte)
+			indexEnd = indexStart + 4 // streamID is 4 bytes
+			streamID := binary.BigEndian.Uint32(headerFrame[indexStart:indexEnd])
+			Expect(streamID).To(BeEquivalentTo(dataStream.StreamID()))
+			// PromisedID
+			indexStart = indexEnd
+			indexEnd = indexStart + 4 // promisedID is 4 bytes
+			promisedID := binary.BigEndian.Uint32(headerFrame[indexStart:indexEnd])
+			Expect(promisedID).To(BeEquivalentTo(pushStream.id))
+		})
+
+		It("Sends a valid header in the PUSH_PROMISE", func() {
+			err := w.Push(pushTarget, opts)
+			Expect(err).ToNot(HaveOccurred())
+			fields := decodePushPromiseFields(headerStream.dataWritten.Bytes())
+			Expect(fields[":method"][0]).To(Equal(method))
+			Expect(fields[":authority"][0]).To(Equal(requestHost))
+			Expect(fields[":path"][0]).To(Equal(pushTarget))
+			Expect(fields[":scheme"][0]).To(Equal("https"))
+		})
+
+		It("pushes", func() {
+			err := w.Push(pushTarget, opts)
+			Expect(err).ToNot(HaveOccurred())
+			Expect(pushStream.dataWritten.Bytes()).To(Equal([]byte(fakePushData)))
+			Expect(pushStream.closed).To(BeTrue())
+		})
+
+		It("pushes targets with a scheme", func() {
+			pushTarget = "https://www.push.com/push_example"
+			err := w.Push(pushTarget, opts)
+			Expect(err).ToNot(HaveOccurred())
+			Expect(pushStream.dataWritten.Bytes()).To(Equal([]byte(fakePushData)))
+			Expect(pushStream.closed).To(BeTrue())
+		})
+
+		It("Does not push when there is no available stream", func() {
+			session.streamsToOpen = []quic.Stream{}
+			err := w.Push(pushTarget, opts)
+			Expect(err).To(HaveOccurred())
+		})
+
+		It("Does not send push_promise on a server initiated stream", func() {
+			headerStream.id = 4
+			err := w.Push(pushTarget, opts)
+			Expect(err).To(HaveOccurred())
+		})
+
+		It("Does not send push resource on a client initiated stream", func() {
+			pushStream.id = 7
+			err := w.Push(pushTarget, opts)
+			Expect(err).To(HaveOccurred())
+		})
+	})
 })
+
+func decodePushPromiseFields(headerData []byte) map[string][]string {
+	fields := make(map[string][]string)
+	decoder := hpack.NewDecoder(4096, func(hf hpack.HeaderField) {})
+	h2framer := http2.NewFramer(nil, bytes.NewReader(headerData))
+
+	frame, err := h2framer.ReadFrame()
+	Expect(err).ToNot(HaveOccurred())
+	Expect(frame).To(BeAssignableToTypeOf(&http2.PushPromiseFrame{}))
+
+	hframe := frame.(*http2.PushPromiseFrame)
+	headerFields, err := decoder.DecodeFull(hframe.HeaderBlockFragment())
+	Expect(err).ToNot(HaveOccurred())
+	for _, p := range headerFields {
+		fields[p.Name] = append(fields[p.Name], p.Value)
+	}
+	return fields
+}