+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## v0.6.0 (unreleased)
 
+- Added HTTP2 Server Push
 - Added `quic.Config` options for maximal flow control windows
 - Add a `quic.Config` option for QUIC versions
 - Add a `quic.Config` option to request truncation of the connection ID from a server